refactor: 본사·창고 영역 폴더 구조 정리 및 미사용 페이지 제거
src/stores 가 권한군별 폴더(warehouse/, store/, hq/circularStock/)
로 분리됐지만 src/api 와 src/views 는 평면 구조로 남아 있어 컨벤션
이 어긋났다. 본인 담당 도메인을 동일한 권한군 구조로 맞추기 위해
api 5개, view 5개 파일을 hq/, warehouse/ 하위 도메인 폴더로 이동
했다.

함께 도메인 적합도 검토 후 제외된 본사 리포트 페이지(사이드바·라
우트·view·store)와 시스템 권장 발주 모달용으로 만들었다 폐기된
aiOrderRecommendation store 를 제거했다.
