+++ src/views/hq/HqProductCreateView.vue
@@ -6,7 +6,7 @@ import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
 import { useAuthStore } from '@/stores/auth.js'
 import { getCategories } from '@/api/category.js'
-import { vendorApi } from '@/api/vendor.js'
+import { vendorApi } from '@/api/hq/vendor.js'
 import {
   createProduct,
   createProductSku,
