+++ src/views/hq/HqProductManagementView.vue
@@ -5,7 +5,7 @@ import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
 import { useAuthStore } from '@/stores/auth.js'
 import { deleteCategory, getCategories, getCategory, updateCategory } from '@/api/category.js'
-import { vendorApi } from '@/api/vendor.js'
+import { vendorApi } from '@/api/hq/vendor.js'
 import {
   getProducts,
 } from '@/api/productMaster.js'
